Broke down code further and respond to comments
