+++ package.json
@@ -29,6 +29,7 @@
 		"mongoose": "^5.9.6",
 		"sirv-cli": "^0.4.4",
 		"socket.io": "^2.3.0",
+		"socket.io-client": "^2.3.0",
 		"typescript": "^3.8.3"
 	}
 }
